DifferentiableKernelFunctions

Documentation for DifferentiableKernelFunctions.

DifferentiableKernelFunctions.EnableDiff

DifferentiableKernelFunctions.apply_partial

DifferentiableKernelFunctions.diffKernelCall

DifferentiableKernelFunctions.diffKernelCall

DifferentiableKernelFunctions.lazy_flatten

DifferentiableKernelFunctions.tangentCurve

DifferentiableKernelFunctions.EnableDiff — Type

EnableDiff

A thin wrapper around Kernels enabling the machinery which allows you to input (∂ᵢ, x), (∂ⱼ, y) where ∂ᵢ, ∂ⱼ are of Partial type (see partial) in order to calculate $k((∂ᵢ, x), (∂ⱼ, y)) = \text{Cov}(\partial_i Z(x), \partial_j Z(y))$ for $Z$ with $k(x,y) = \text{Cov}(Z(x), Z(y))$.

!!! warning Only apply this wrapper at the very end. Kerneltransformations should be applied beforehand.

!!! info While this machinery could in principle be enabled for all Kernel by default, the covariance of derivatives of an isotropic kernel are no longer isotropic. This forces the use of less specialized methods. So for now you have to opt-in with this Wrapper.

Example:

julia> k = EnableDiff(SEKernel());

julia> k((0, partial(1)), 0) # calculate Cov(∂₁Z(0), Z(0))
0.0

julia> k(0,0) # normal input still works
1.0

DifferentiableKernelFunctions.apply_partial — Method

Take the partial derivative of a function with two dim-dimensional inputs, i.e. 2*dim dimensional input

DifferentiableKernelFunctions.diffKernelCall — Method

Multi Kernel Version (do not try to take the derivative with regard to out indices)

DifferentiableKernelFunctions.diffKernelCall — Method

diffKernelCall(k::T, (px,x)::DiffPt, (py, y)::DiffPt) where {Dim, T<:Kernel}

specialization for DiffPt. Unboxes the partial instructions from DiffPt and applies them to k, evaluates them at the positions of DiffPt

DifferentiableKernelFunctions.lazy_flatten — Method

lazy_flatten(vectors...)

The output is a lazy form of

collect(Iterators.flatten(vectors...))

i.e. it is an AbstractArray in contrast to Iterators.flatten(vectors...). So is accessible with getindex and gets default Array implementations for free. In particular it can be passed to Base.PermutedDimsArrayfor lazy permutation andvec()to obtain a lazyBase.ReshapedArray`.

DifferentiableKernelFunctions.tangentCurve — Method

tangentCurve(x₀, i::IndexType)

returns the function (t ↦ x₀ + teᵢ) where eᵢ is the unit vector at index i