DifferentiableKernelFunctions

Documentation for DifferentiableKernelFunctions.

DifferentiableKernelFunctions._evaluate

DifferentiableKernelFunctions.apply_partial

DifferentiableKernelFunctions.lazy_flatten

DifferentiableKernelFunctions.lazy_product

DifferentiableKernelFunctions.tangentCurve

DifferentiableKernelFunctions._evaluate — Method

_evaluate(k::T, x::DiffPt{Dim}, y::DiffPt{Dim}) where {Dim, T<:Kernel}

implements (k::T)(x::DiffPt{Dim}, y::DiffPt{Dim}) for all kernel types. But since generics are not allowed in the syntax above by the dispatch system, this redirection over _evaluate is necessary

unboxes the partial instructions from DiffPt and applies them to k, evaluates them at the positions of DiffPt

DifferentiableKernelFunctions.apply_partial — Method

Take the partial derivative of a function with two dim-dimensional inputs, i.e. 2*dim dimensional input

DifferentiableKernelFunctions.lazy_flatten — Method

lazy_flatten(vectors...)

The output is a lazy form of

collect(Iterators.flatten(vectors...))

i.e. it is an AbstractArray in contrast to Iterators.flatten(vectors...). So is accessible with getindex and gets default Array implementations for free. In particular it can be passed to Base.PermutedDimsArrayfor lazy permutation andvec()to obtain a lazyBase.ReshapedArray`.

DifferentiableKernelFunctions.lazy_product — Method

lazy_product(vectors...)

The output is a lazy form of

collect(Iterators.product(vectors...))

i.e. it is an AbstractArray in contrast to Iterators.product(vectors...). So is accessible with getindex and gets default Array implementations for free. In particular it can be passed to Base.PermutedDimsArrayfor lazy permutation andvec()to obtain a lazyBase.ReshapedArray`.

DifferentiableKernelFunctions.tangentCurve — Method

tangentCurve(x₀, i::IndexType)

returns the function (t ↦ x₀ + teᵢ) where eᵢ is the unit vector at index i